+++ Runtime/Pbr/Manipulators/RotationProxy.cs
@@ -3,15 +3,16 @@ using UnityEngine.UIElements;
 
 namespace Unity.Muse.Texture
 {
-    internal class RotationProxy: Manipulator 
+    internal class RotationProxy : Manipulator
     {
+        public bool active { get; set; } = true;
         RotationManipulator m_RotationManipulator;
-        
+
         public RotationProxy(RotationManipulator rotationManipulator)
         {
             m_RotationManipulator = rotationManipulator;
         }
-        
+
         protected override void RegisterCallbacksOnTarget()
         {
             target.RegisterCallback<PointerDownEvent>(OnPointerDown);
@@ -19,7 +20,7 @@ namespace Unity.Muse.Texture
             target.RegisterCallback<PointerUpEvent>(OnPointerUp);
             target.RegisterCallback<MouseMoveEvent>(OnMouseMove);
         }
-        
+
         protected override void UnregisterCallbacksFromTarget()
         {
             target.UnregisterCallback<PointerDownEvent>(OnPointerDown);
@@ -30,27 +31,35 @@ namespace Unity.Muse.Texture
 
         private void OnMouseMove(MouseMoveEvent evt)
         {
+            if (!active)
+                return;
+
             RotationManipulator.OnMouseMove(evt, false);
         }
 
         private void OnPointerUp(PointerUpEvent evt)
         {
+            if (!active)
+                return;
+            
             target.ReleasePointer(evt.pointerId);
-           m_RotationManipulator.OnPointerUp(evt); 
+            m_RotationManipulator.OnPointerUp(evt);
         }
 
         private void OnPointerMove(PointerMoveEvent evt)
         {
-            m_RotationManipulator.OnPointerMove(evt, false); 
+            if(!active)
+                return;
+            m_RotationManipulator.OnPointerMove(evt, false, false);
         }
 
         private void OnPointerDown(PointerDownEvent evt)
         {
-            if(evt.button != 0)
+            if (!active || evt.button != 0)
                 return;
-            
+
             target.CapturePointer(evt.pointerId);
-            m_RotationManipulator.OnPointerDown(evt, false, false); 
+            m_RotationManipulator.OnPointerDown(evt, false, false);
         }
     }
 }