+++ CHANGELOG.md
@@ -4,6 +4,26 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-pre.18] - 2024-03-11
+
+## [1.0.0-pre.5] - 2024-02-16
+
+### Changed
+
+- Update com.unity.muse.common dependency version.
+
+## [1.0.0-pre.4] - 2024-02-16
+
+### Changed
+
+- Prompt fields are now pre-populated when going into refine mode.
+- Prompt fields are now disabled when using in-painting.
+
+### Fixed
+
+- Fix Generations label overflowing.
+- Fix Material preview rotation that was resetting.
+
 ## [1.0.0-pre.3] - 2023-12-15
 
 ### Fixed